upgrade: 前端页面基本完成, 后端大部分BUG已修复, 仍存在回合数切换问题

+++ src/main/java/com/szm/demo/mapper/PlayerActionGroupMapper.java
@@ -11,5 +11,5 @@ import java.util.List;
 public interface PlayerActionGroupMapper {
 
     @Select("SELECT * FROM player_action_group WHERE level_id=#{levelId}")
-    List<PlayerActionGroup> getByLId(@Param("levelId") Long levelId);
+    List<PlayerActionGroup> getByLId(@Param("levelId") Integer levelId);
 }
